tests/integration/test_server_errors.py
# tests/integration/test_server_errors.py
import pytest
import requests
from unittest.mock import patch, AsyncMock
from .conftest import BASE_URL, valid_headers, minimal_chat_payload, minimal_embedding_payload, configured_chat_model_id, configured_embedding_model_id

@pytest.mark.asyncio
@patch('app.routers.api_v1.openai_chat_request_to_core') # Patching a function within the router
async def test_ecv_server_001_unhandled_exception_chat(mock_conversion, valid_headers, minimal_chat_payload):
    """ECV_SERVER_001: Simulate an unhandled Python error in /chat/completions endpoint logic."""
    mock_conversion.side_effect = Exception("Simulated generic unhandled error")
    response = requests.post(f"{BASE_URL}/chat/completions", json=minimal_chat_payload, headers=valid_headers)
    assert response.status_code == 500
    json_response = response.json()
    assert json_response["detail"] == "Internal Server Error"
    assert "request_id" in json_response # From StructlogMiddleware

@pytest.mark.asyncio
@patch('app.auth.repositories.APIKeyRepository.get_by_api_key_value')
async def test_ecv_server_002_db_connection_failure_auth(mock_get_key, valid_headers):
    """ECV_SERVER_002: Simulate DB connection failure during API key validation."""
    from sqlalchemy.exc import OperationalError # Import a relevant SQLAlchemy exception
    mock_get_key.side_effect = OperationalError("Simulated DB connection failure", params=None, orig=None)
    
    # Any endpoint that uses valid_api_key dependency
    response = requests.get(f"{BASE_URL}/models", headers=valid_headers)
    assert response.status_code == 500
    json_response = response.json()
    assert json_response["detail"] == "Internal Server Error"
    assert "request_id" in json_response

@pytest.mark.asyncio
@patch('app.providers.bedrock.bedrock.aioboto3.Session')
async def test_ecv_server_004_downstream_bedrock_api_error(mock_aioboto_session, valid_headers, configured_chat_model_id):
    """ECV_SERVER_004: Simulate Bedrock client API returning an error."""
    from botocore.exceptions import ClientError

    mock_client = AsyncMock()
    # Simulate a ClientError from Bedrock
    mock_client.converse.side_effect = ClientError(
        error_response={'Error': {'Code': 'ValidationException', 'Message': 'Mocked Bedrock ClientError'}},
        operation_name='Converse'
    )
    mock_aioboto_session.return_value.client.return_value.__aenter__.return_value = mock_client
    mock_aioboto_session.return_value.client.return_value.__aexit__.return_value = None

    payload = {"model": configured_chat_model_id, "messages": [{"role": "user", "content": "Hello"}]}
    # Ensure configured_chat_model_id is one handled by BedRockBackend in your settings
    if "claude" not in configured_chat_model_id and "llama" not in configured_chat_model_id: # Example check
        pytest.skip(f"Skipping Bedrock specific test for model {configured_chat_model_id}")

    response = requests.post(f"{BASE_URL}/chat/completions", json=payload, headers=valid_headers)
    assert response.status_code == 500 # As current code doesn't specifically translate provider errors
    json_response = response.json()
    assert json_response["detail"] == "Internal Server Error"
    assert "request_id" in json_response

@pytest.mark.asyncio
@patch('app.providers.vertex_ai.vertexai.GenerativeModel')
async def test_ecv_server_005_downstream_vertex_api_error(MockGenerativeModel, valid_headers, configured_chat_model_id):
    """ECV_SERVER_005: Simulate Vertex AI client API raising an exception."""
    from google.api_core.exceptions import GoogleAPIError

    mock_model_instance = AsyncMock()
    mock_model_instance.generate_content_async.side_effect = GoogleAPIError("Simulated Vertex GoogleAPIError")
    MockGenerativeModel.return_value = mock_model_instance

    payload = {"model": configured_chat_model_id, "messages": [{"role": "user", "content": "Hello"}]}
    # Ensure configured_chat_model_id is one handled by VertexBackend in your settings
    if "gemini" not in configured_chat_model_id: # Example check
         pytest.skip(f"Skipping Vertex specific test for model {configured_chat_model_id}")

    response = requests.post(f"{BASE_URL}/chat/completions", json=payload, headers=valid_headers)
    assert response.status_code == 500
    json_response = response.json()
    assert json_response["detail"] == "Internal Server Error"
    assert "request_id" in json_response

# Note: ECV_SERVER_006, ECV_SERVER_007 (Timeouts/Network Issues) are harder to reliably simulate
# without more complex network mocking tools or specific provider SDK timeout exceptions.
# They would typically also result in a 500 if the SDK raises an unhandled exception.

# Note: ECV_SERVER_008 (Billing Service Queue Failure) is conceptual.
# If billing_queue.put() was to raise an unhandled error, it would lead to a 500.
# Example:
@pytest.mark.asyncio
@patch('app.services.billing.billing_queue.put')
async def test_ecv_server_008_billing_queue_failure(mock_billing_put, valid_headers, minimal_chat_payload):
    """ECV_SERVER_008: Simulate billing queue put operation failing."""
    mock_billing_put.side_effect = Exception("Simulated billing queue error")
    
    # This assumes billing_queue.put is called after a successful model response.
    # To test this properly, we might need to mock the actual model call to succeed first,
    # then have billing fail. For simplicity, we'll assume it's called.
    # A more robust test would involve patching the model provider to return a valid response,
    # then checking if the billing error is handled or results in 500.
    
    # For this example, let's assume the endpoint logic is:
    # try:
    #   model_response = await backend.invoke_model(...)
    #   await billing_queue.put(...) # This is where we mock the error
    #   return ...
    # except Exception: -> 500
    
    # To make this testable, we need a successful path that then hits the billing.
    # We'll patch the actual model invocation to succeed, then cause billing to fail.
    with patch('app.providers.base.Backend.invoke_model', new_callable=AsyncMock) as mock_invoke:
        from app.providers.core.chat_schema import ChatRepsonse, Response as CoreResponse, CompletionUsage
        from datetime import datetime

        mock_invoke.return_value = ChatRepsonse(
            created=datetime.now(), model=minimal_chat_payload["model"],
            choices=[CoreResponse(content="Successful mock response")],
            usage=CompletionUsage(prompt_tokens=1, completion_tokens=1, total_tokens=2)
        )

        response = requests.post(f"{BASE_URL}/chat/completions", json=minimal_chat_payload, headers=valid_headers)
        
        # If the billing error is unhandled within the endpoint, it should result in a 500.
        # If the endpoint handles the billing error gracefully (e.g., logs it but still returns 200),
        # this assertion would change.
        assert response.status_code == 500
        json_response = response.json()
        assert json_response["detail"] == "Internal Server Error"
        assert "request_id" in json_response
        mock_billing_put.assert_called_once()


# ECV_SERVER_009 and ECV_SERVER_010 are about application lifespan events (shutdown)
# and are not typically tested via HTTP requests to endpoints. They would require
# different testing strategies, possibly involving direct calls to the lifespan
# manager or instrumenting the application shutdown process.

