requirements.txt
# Generated by setup_macos.sh
aioboto3>=14.1.0
alembic>=1.15.1
fastapi>=0.115.11
google-cloud-aiplatform>=1.90.0
# google-generativeai # Removed as vertexai is primary for this project
greenlet>=3.1.1
psycopg[binary]>=3.2.5 # For postgresql+psycopg connection string
pydantic>=2.10.6
pydantic-settings>=2.8.1
sqlalchemy>=2.0.38
structlog>=25.2.0
types-aioboto3>=14.1.0
uvicorn>=0.34.0
# vertexai is part of google-cloud-aiplatform

# Test dependencies
pytest>=8.3.5
pytest-asyncio>=0.26.0
pytest-mock>=3.14.0

# Linter
ruff>=0.9.10
